requirements.txt
streamlit>=1.35.0
openai>=1.0.0
langchain>=0.2.0
python-dotenv>=1.0.1